delete GameStateFilter
