+++ crates/control/src/game_controller_state_filter.rs
@@ -7,14 +7,10 @@ use nalgebra::{distance, Isometry2, Point2, Vector2};
 use serde::{Deserialize, Serialize};
 use spl_network_messages::{GamePhase, GameState, Team};
 use types::{
-    ball_position::BallPosition,
-    cycle_time::CycleTime,
-    field_dimensions::FieldDimensions,
+    ball_position::BallPosition, cycle_time::CycleTime, field_dimensions::FieldDimensions,
     filtered_game_controller_state::FilteredGameControllerState,
-    filtered_game_states::FilteredGameState,
-    filtered_whistle::FilteredWhistle,
-    game_controller_state::GameControllerState,
-    parameters::GameStateFilterParameters,
+    filtered_game_states::FilteredGameState, filtered_whistle::FilteredWhistle,
+    game_controller_state::GameControllerState, parameters::GameStateFilterParameters,
 };
 #[derive(Deserialize, Serialize)]
 pub struct GameControllerStateFilter {
@@ -159,23 +155,43 @@ fn next_filtered_state(
     ball_detected_far_from_any_goal: bool,
 ) -> State {
     match (current_state, game_controller_state.game_state) {
-        (State::Initial | State::Ready | State::Finished, _)
-        | (
-            State::Set,
-            GameState::Initial | GameState::Ready | GameState::Playing | GameState::Finished,
-        )
+        (State::Finished, _) => State::Finished,
+
+        (
+            State::TentativeFinished {
+                time_when_finished_clicked,
+            },
+            GameState::Finished,
+        ) if cycle_start_time
+            .duration_since(time_when_finished_clicked)
+            .unwrap()
+            >= config.tentative_finish_duration =>
+        {
+            State::Finished
+        }
+        (
+            State::TentativeFinished {
+                time_when_finished_clicked,
+            },
+            GameState::Finished,
+        ) => State::TentativeFinished {
+            time_when_finished_clicked,
+        },
+        (State::TentativeFinished { .. }, game_state) => State::from_game_state(game_state),
+        (_, GameState::Finished) => State::TentativeFinished {
+            time_when_finished_clicked: cycle_start_time,
+        },
+        (State::Initial | State::Ready, _)
+        | (State::Set, GameState::Initial | GameState::Ready | GameState::Playing)
         | (
             State::WhistleInSet { .. },
-            GameState::Initial | GameState::Ready | GameState::Playing | GameState::Finished,
+            GameState::Initial | GameState::Ready | GameState::Playing,
         )
-        | (
-            State::Playing,
-            GameState::Initial | GameState::Ready | GameState::Set | GameState::Finished,
-        )
+        | (State::Playing, GameState::Initial | GameState::Ready | GameState::Set)
         | (
             State::WhistleInPlaying { .. },
-            GameState::Initial | GameState::Ready | GameState::Set | GameState::Finished,
-        ) => State::from_game_controller(game_controller_state),
+            GameState::Initial | GameState::Ready | GameState::Set,
+        ) => State::from_game_state(game_controller_state.game_state),
         (State::Set, GameState::Set) => {
             if is_whistle_detected {
                 State::WhistleInSet {
@@ -274,12 +290,15 @@ pub enum State {
     WhistleInPlaying {
         time_when_whistle_was_detected: SystemTime,
     },
+    TentativeFinished {
+        time_when_finished_clicked: SystemTime,
+    },
     Finished,
 }
 
 impl State {
-    fn from_game_controller(game_controller_state: &GameControllerState) -> Self {
-        match game_controller_state.game_state {
+    fn from_game_state(game_state: GameState) -> Self {
+        match game_state {
             GameState::Initial => State::Initial,
             GameState::Ready => State::Ready,
             GameState::Set => State::Set,
@@ -369,6 +388,8 @@ impl State {
                 GamePhase::PenaltyShootout { .. } => FilteredGameState::Set,
                 _ => FilteredGameState::Finished,
             },
+            // is hack @schluis
+            State::TentativeFinished { .. } => FilteredGameState::Set,
         }
     }
 }
